+++ Ecolocation/Ecolocation/app/src/main/res/layout/activity_google_drive.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="ecolocation.ecolocation.GoogleDriveDemoActivity">
+    tools:context="ecolocation.ecolocation.ListViewActivity">
 
     <TextView
         android:id="@+id/txt_list_title"
