+++ Ecolocation/app/src/main/java/ecolocation/ecolocation/AnimalDetailActivity.java
@@ -3,16 +3,14 @@ package ecolocation.ecolocation;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.squareup.picasso.Picasso;
-
-import java.util.ArrayList;
-
-import java.util.ArrayList;
 
 public class AnimalDetailActivity extends AppCompatActivity {
     //widgets
@@ -31,14 +29,19 @@ public class AnimalDetailActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_animal_detail);
 
-        //ArrayList<Animal> animalList = (ArrayList<Animal>) getIntent().getSerializableExtra("animal_list");
+        //----------- toolbar setup
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        toolbar.setNavigationIcon(R.drawable.ic_up_navigation);
+
         back_button = (Button) findViewById(R.id.back_button);
 
         back_button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 //TODO: reset to original values
-                Intent intent = new Intent( AnimalDetailActivity.this, GoogleDriveDemoActivity.class );
+                Intent intent = new Intent( AnimalDetailActivity.this, ListViewActivity.class );
                 startActivity( intent );
                 finish();
             }
@@ -58,5 +61,17 @@ public class AnimalDetailActivity extends AppCompatActivity {
 
 
         Picasso.with(this).load(url).error(R.mipmap.ic_launcher).into(animalPic);
+    }
+
+    //used to go back to the ListView Activity
+    public boolean onOptionsItemSelected(MenuItem item) {
+        //go to ListView Activity
+        switch (item.getItemId()) {
+            case android.R.id.home:
+                Intent intent = new Intent(AnimalDetailActivity.this, ListViewActivity.class);
+                startActivity(intent);
+                break;
+        }
+        return true;
     }
 }
