+++ Ecolocation/app/src/main/java/ecolocation/ecolocation/ListViewActivity.java
@@ -1,20 +1,39 @@
 package ecolocation.ecolocation;
 
+import android.annotation.SuppressLint;
+import android.content.Intent;
 import android.graphics.drawable.Drawable;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
+import android.widget.ImageView;
 import android.widget.ListView;
 
+import com.squareup.picasso.Picasso;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.IOException;
 import java.util.ArrayList;
 
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
+
 public class ListViewActivity extends AppCompatActivity {
-    //widgets
+    //Widgets
     ListView listView;
     Button resetButton;
     Button recalcButton;
     Button sortButton;
+    static int flag = 0;
+
 
     //variables for creating the list
     private ArrayList<Animal> animalList;
@@ -24,6 +43,12 @@ public class ListViewActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_list_view);
+
+        //----------- toolbar setup
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        toolbar.setNavigationIcon(R.drawable.ic_up_navigation);
 
         //TODO: initialize animalList with database/google drive stuff;
         animalList = fillList();
@@ -51,13 +76,20 @@ public class ListViewActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 //TODO: reset to original values
+                Intent intent = new Intent( ListViewActivity.this, ListViewActivity.class );
+                startActivity( intent );
+                finish();
             }
         });
 
+        // go to detailed page
         recalcButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                //TODO: calculate new nutrient dispersal
+                Intent intent = new Intent( ListViewActivity.this, AnimalDetailActivity.class );
+                //intent.putExtra("animal_list", animalList);
+                startActivity( intent );
+                finish();
             }
         });
 
@@ -67,34 +99,101 @@ public class ListViewActivity extends AppCompatActivity {
                 //TODO: show different ways of sorting
             }
         });
+
     }
 
     //this is just filling it in with dummy data
     private ArrayList<Animal> fillList(){
-        Drawable pic = getResources().getDrawable(R.drawable.ic_launcher_background);
 
-        Animal lion = new Animal("Panthera Leo","lion", pic,"A big cat in Africa", "carnivore",
-                "vulnerable", 187.5, 20000);
+        // default image to display in case something happens
+        final Drawable pic = getResources().getDrawable(R.drawable.ic_launcher_background);
+        final ArrayList<Animal> list = new ArrayList<>();
 
-        Animal elephant = new Animal("Loxodonta africana","african elephant", pic, "The largest land mammal",
-                "herbivore", "vulnerable", 3500, 415000);
+        @SuppressLint("StaticFieldLeak") AsyncTask<Integer, Void, Void> asyncTask = new AsyncTask<Integer, Void, Void>() {
+            @Override
+            protected Void doInBackground(Integer... Void) {
 
-        Animal giraffe = new Animal("Giraffa camelopardalis","giraffe", pic, "An animal with a long neck",
-                "herbivore", "vulnerable", 1192, 97500);
+                OkHttpClient client = new OkHttpClient();
+                // animals.php is old db call for just getting binomial
+                /*Request request = new Request.Builder()
+                        .url("http://18.220.129.239/animals.php?")
+                        .build();*/
+                Request request = new Request.Builder()
+                        .url("http://18.220.129.239/mammals.php?")
+                        .build();
+                try {
+                    Response response = client.newCall(request).execute();
 
-        Animal cheetah = new Animal("Acinonyx jubatus", "cheetah", pic, "A very fast animal",
-                "carnivore", "vulnerable", 50, 7100);
+                    JSONArray array = new JSONArray(response.body().string());
 
-        Animal zebra = new Animal("Equus zebra","zebra", pic, "A striped horse.", "herbivore",
-                "near threatened", 250, 150000);
+                    for (int i = 0; i < array.length(); i++) {
 
-        ArrayList<Animal> list = new ArrayList<Animal>();
-        list.add(lion);
-        list.add(elephant);
-        list.add(giraffe);
-        list.add(cheetah);
-        list.add(zebra);
+                        JSONObject object = array.getJSONObject(i);
+
+                        Animal animal = new Animal(object.getString("binomial"), object.getString("common_name"), pic,
+                                "A big cat in Africa", "Carnivore", object.getString("endangered_level"),
+                                object.getInt("mass"), object.getInt("population"));
+
+
+                        //ListViewActivity.this.animalList.add(animal);
+                        list.add(animal);
+                        Log.d("return", animal.getBinomial());
+                    }
+
+
+
+                } catch (IOException e) {
+                    e.printStackTrace();
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+                return null;
+            }
+
+            @Override
+            protected void onPostExecute(Void aVoid) {
+                adapter.notifyDataSetChanged();
+                for(int i=0; i<animalList.size(); i++) {
+                    Animal currAnimal = animalList.get(i);
+                    loadImageFromURL(currAnimal);
+                    Log.d("currAnimal", currAnimal.getBinomial());
+                    // Do something with the value
+                }
+            }
+        };
+
+        asyncTask.execute();
 
         return list;
+    }
+
+
+
+    // function to load an image into an image view
+    private void loadImageFromURL(final Animal animal)
+    {
+        // create an imageView to hold the picture
+        final ImageView imageView = new ImageView(this);
+        String url = "http://cefns.nau.edu/~mh973/images/" + animal.getName() + ".jpg";
+        // call to get picture
+        Picasso.with(this).load(url).error(R.mipmap.ic_launcher).into(imageView, new com.squareup.picasso.Callback(){
+
+            // because the image doesn't load all at once you have to set the image for the animal when it is successful
+            @Override
+            public void onSuccess()
+            {
+                Drawable d = imageView.getDrawable();
+                animal.setImage(d);
+                if( flag == 0 )
+                {
+                    Intent intent = new Intent( ListViewActivity.this, ListViewActivity.class );
+                    startActivity( intent );
+                    finish();
+                    flag = 1;
+                }
+            }
+            @Override
+            public void onError(){}
+        });
     }
 }
