--- Ecolocation/app/src/main/res/layout/activity_google_drive.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context="ecolocation.ecolocation.GoogleDriveDemoActivity">
-
-    <TextView
-        android:id="@+id/txt_list_title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_margin="20dp"
-        android:text="Demo of the Google Drive Data Retrieval"
-        android:textAppearance="@style/TextAppearance.AppCompat.Medium" />
-
-    <!-- The actual list, where each item has the layout of list_item.xml -->
-    <ListView
-        android:id="@+id/layout_list"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/bttn_recalc"
-        android:layout_below="@+id/txt_list_title"
-        android:layout_gravity="center" />
-
-    <Button
-        android:id="@+id/bttn_reset"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="false"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="false"
-        android:layout_gravity="center"
-        android:text="@string/bttn_reset" />
-
-    <Button
-        android:id="@+id/bttn_recalc"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_gravity="center"
-        android:text="@string/bttn_re_run" />
-
-    <Button
-        android:id="@+id/bttn_sort"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentLeft="true"
-        android:layout_gravity="center"
-        android:text="@string/bttn_sort" />
-
-
-</RelativeLayout>
