+++ Ecolocation/app/src/main/res/layout/activity_info.xml
@@ -7,10 +7,17 @@
     android:layout_height="match_parent"
     tools:context="ecolocation.ecolocation.InfoActivity">
 
+    <include
+        android:id="@+id/toolbar"
+        layout="@layout/toolbar_main"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
+
     <TextView
         android:id="@+id/txt_info_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_below="@+id/toolbar"
         android:layout_marginTop="30dp"
         android:padding="10dp"
         android:text="@string/info_str"
