+++ Ecolocation/app/src/main/res/layout/activity_list_view.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -8,13 +7,20 @@
     android:orientation="vertical"
     tools:context="ecolocation.ecolocation.ListViewActivity">
 
+    <include
+        android:id="@+id/toolbar"
+        layout="@layout/toolbar_main"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
 
     <TextView
         android:id="@+id/txt_list_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_margin="20dp"
+        android:layout_below="@+id/toolbar"
+        android:layout_centerHorizontal="true"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginTop="18dp"
         android:text="@string/list_title_str"
         android:textAppearance="@style/TextAppearance.AppCompat.Medium" />
 
@@ -25,8 +31,7 @@
         android:layout_height="wrap_content"
         android:layout_above="@+id/bttn_recalc"
         android:layout_below="@+id/txt_list_title"
-        android:layout_gravity="center"
-        android:layout_marginBottom="10dp" />
+        android:layout_gravity="center" />
 
     <Button
         android:id="@+id/bttn_reset"
@@ -35,6 +40,7 @@
         android:layout_alignParentBottom="true"
         android:layout_alignParentRight="false"
         android:layout_centerHorizontal="true"
+        android:layout_centerVertical="false"
         android:layout_gravity="center"
         android:text="@string/bttn_reset" />
 
@@ -45,7 +51,7 @@
         android:layout_alignParentBottom="true"
         android:layout_alignParentRight="true"
         android:layout_gravity="center"
-        android:text="@string/bttn_run" />
+        android:text="@string/bttn_re_run" />
 
     <Button
         android:id="@+id/bttn_sort"
